refactor: Align package name with repo and remove src layout
